+++ src/expr/sql.rs
@@ -180,7 +180,9 @@ impl<'a> expr::Visitor<'a, ast::Expr> for FromExprVisitor {
             | expr::function::Function::Random(_)
             | expr::function::Function::Least
             | expr::function::Function::Greatest
-            | expr::function::Function::Coalesce => ast::Expr::Function(ast::Function {
+            | expr::function::Function::Coalesce
+            | expr::function::Function::Rtrim
+            | expr::function::Function::Ltrim => ast::Expr::Function(ast::Function {
                 name: ast::ObjectName(vec![ast::Ident::new(function.to_string())]),
                 args: arguments
                     .into_iter()